introspectMethods dans Utils
